Move to gaussian_process sub-module
