+++ skopt/learning/gaussian_process/__init__.py
@@ -1,0 +1,3 @@
+from .gpr import GaussianProcessRegressor
+
+__all__ = ("GaussianProcessRegressor")
