+++ skopt/learning/gaussian_process/kernels.py
